+++ example/lib/main.dart
@@ -50,7 +50,7 @@ class _MyHomePageState extends State<MyHomePage> {
               padding: EdgeInsets.symmetric(horizontal: 20),
               child: Container(
                   alignment: Alignment.centerLeft,
-                  child: Text("Gradient Borders",
+                  child: Text("Snake Border",
                       style: TextStyle(
                           color: Colors.white, fontWeight: FontWeight.bold))),
             ),
@@ -58,218 +58,52 @@ class _MyHomePageState extends State<MyHomePage> {
               height: 20,
             ),
             SizedBox(
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: [
-                  ZoAnimatedGradientBorder(
-                    borderThickness: 8,
-                    borderRadius: 10,
-                    glowOpacity: 1.0,
-                    gradientColor: [Colors.yellow, Colors.orange],
-                    duration: Duration(seconds: 4),
+                child: Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                    children: [
+                  ZoSnakeBorder(
+                    duration: 3,
+                    glowOpacity: 0,
+                    snakeHeadColor: Colors.red,
+                    snakeTailColor: Colors.blue,
+                    snakeTrackColor: Colors.blueGrey,
+                    borderWidth: 5,
+                    borderRadius: BorderRadius.circular(10),
                     child: Container(
-                        width: 100,
-                        height: 100,
-                        alignment: Alignment.center,
-                        decoration: BoxDecoration(
-                            color: Colors.red,
-                            borderRadius: BorderRadius.circular(10)),
-                        child: Text(
-                          "Color",
-                          style: TextStyle(color: Colors.black),
-                        )),
+                      width: 150,
+                      height: 45,
+                      alignment: Alignment.center,
+                      decoration: BoxDecoration(
+                          color: Colors.white,
+                          borderRadius: BorderRadius.circular(10)),
+                      child: Text(
+                        "Snake Border",
+                        style: TextStyle(color: Colors.black),
+                      ),
+                    ),
                   ),
-                  ZoAnimatedGradientBorder(
-                    borderRadius: 100,
-                    borderThickness: 4,
-                    gradientColor: [Colors.orange, Colors.white, Colors.green],
-                    duration: Duration(seconds: 4),
+                  ZoSnakeBorder(
+                    duration: 3,
+                    glowOpacity: 0.3,
+                    snakeHeadColor: Colors.red,
+                    snakeTailColor: Colors.blue,
+                    snakeTrackColor: Colors.blueGrey,
+                    borderWidth: 7,
+                    borderRadius: BorderRadius.circular(60),
                     child: Container(
-                        width: 100,
-                        height: 100,
-                        alignment: Alignment.center,
-                        decoration: BoxDecoration(
-                            color: Colors.black,
-                            borderRadius: BorderRadius.circular(50)),
-                        child: Text(
-                          "Color",
-                          style: TextStyle(color: Colors.black),
-                        )),
-                  ),
-                ],
-              ),
-            ),
-            // SizedBox(
-            //   height: 50,
-            // ),
-            // Padding(
-            //   padding: const EdgeInsets.symmetric(horizontal: 10),
-            //   child: Row(
-            //     mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            //     children: [
-            //       ZoAnimatedGradientBorder(
-            //         // width: 150,
-            //         // height: 45,
-            //         borderThickness: 2,
-            //         shouldAnimate: false,
-            //         gradientColor: [Colors.red, Colors.blue],
-            //         duration: Duration(seconds: 4),
-            //         child: Container(
-            //             alignment: Alignment.center,
-            //             color: Colors.black,
-            //             child: Text(
-            //               "Click Me",
-            //               style: TextStyle(color: Colors.white),
-            //             )),
-            //       ),
-            //       ZoAnimatedGradientBorder(
-            //         // width: 150,
-            //         // height: 45,
-            //         borderThickness: 3,
-            //         shouldAnimate: true,
-            //         gradientColor: [Colors.orange, Colors.white, Colors.green],
-            //         duration: Duration(seconds: 4),
-            //         child: Container(
-            //             alignment: Alignment.center,
-            //             color: Colors.white,
-            //             child: Text(
-            //               "Click Me",
-            //               style: TextStyle(color: Colors.black),
-            //             )),
-            //       ),
-            //     ],
-            //   ),
-            // ),
-            // SizedBox(
-            //   height: 25,
-            // ),
-            // Padding(
-            //   padding: const EdgeInsets.all(8.0),
-            //   child: Container(
-            //       alignment: Alignment.centerLeft,
-            //       child: Text("Styled Monocrome Borders",
-            //           style: TextStyle(color: Colors.white))),
-            // ),
-            // SizedBox(
-            //   height: 10,
-            // ),
-            // Padding(
-            //   padding: const EdgeInsets.all(8.0),
-            //   child: Row(
-            //     mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            //     children: [
-            //       ZoMonoCromeBorder(
-            //         trackBorderColor: Colors.white,
-            //         cornerRadius: 50.0,
-            //         borderStyle: ZoMonoCromeBorderStyle.mirror,
-            //         borderWidth: 5.5,
-            //         child: Container(
-            //           width: 100,
-            //           height: 100,
-            //           alignment: Alignment.center,
-            //           child:
-            //               Text("Mirror", style: TextStyle(color: Colors.white)),
-            //           decoration: BoxDecoration(
-            //             color: Colors.red,
-            //             shape: BoxShape.circle,
-            //           ),
-            //         ),
-            //       ),
-            //       ZoMonoCromeBorder(
-            //         trackBorderColor: Colors.white,
-            //         cornerRadius: 50.0,
-            //         borderStyle: ZoMonoCromeBorderStyle.repeated,
-            //         borderWidth: 5.5,
-            //         child: Container(
-            //           width: 100,
-            //           height: 100,
-            //           alignment: Alignment.center,
-            //           child: Text("Repeated",
-            //               style: TextStyle(color: Colors.white)),
-            //           decoration: BoxDecoration(
-            //             color: Colors.red,
-            //             shape: BoxShape.circle,
-            //           ),
-            //         ),
-            //       ),
-            //       ZoMonoCromeBorder(
-            //         trackBorderColor: Colors.white,
-            //         cornerRadius: 50.0,
-            //         borderStyle: ZoMonoCromeBorderStyle.stroke,
-            //         borderWidth: 5.5,
-            //         child: Container(
-            //           width: 100,
-            //           height: 100,
-            //           alignment: Alignment.center,
-            //           child:
-            //               Text("Stroke", style: TextStyle(color: Colors.white)),
-            //           decoration: BoxDecoration(
-            //             color: Colors.red,
-            //             shape: BoxShape.circle,
-            //           ),
-            //         ),
-            //       ),
-            //     ],
-            //   ),
-            // ),
-            // Padding(
-            //   padding: EdgeInsets.symmetric(horizontal: 20),
-            //   child: Container(
-            //       alignment: Alignment.centerLeft,
-            //       child: Text("Pulse Borders",
-            //           style: TextStyle(
-            //               color: Colors.white, fontWeight: FontWeight.bold))),
-            // ),
-            // SizedBox(
-            //   height: 40,
-            // ),
-            // Padding(
-            //   padding: const EdgeInsets.symmetric(horizontal: 40),
-            //   child: Row(
-            //     children: [
-            //       ZoPulsatingBorder(
-            //         layerCount: 2,
-            //         type: ZoPulsatingBorderType.radarPulse,
-            //         pulseColor: Colors.blue,
-            //         borderRadius: BorderRadius.circular(10),
-            //         child: Container(
-            //           alignment: Alignment.center,
-            //           width: 120,
-            //           height: 45,
-            //           decoration: BoxDecoration(
-            //               color: Colors.red,
-            //               borderRadius: BorderRadius.circular(10)),
-            //           child: Text(
-            //             "Radar Pulse",
-            //             style: TextStyle(color: Colors.white),
-            //           ),
-            //         ),
-            //       ),
-            //       SizedBox(
-            //         width: 80,
-            //       ),
-
-            //       // For circle pulse
-            //       ZoPulsatingBorder(
-            //         type: ZoPulsatingBorderType.pulse,
-            //         borderRadius: BorderRadius.circular(100),
-            //         pulseColor: Colors.blue,
-            //         child: Container(
-            //           alignment: Alignment.center,
-            //           width: 100,
-            //           height: 100,
-            //           decoration: BoxDecoration(
-            //               color: Colors.red,
-            //               borderRadius: BorderRadius.circular(50)),
-            //           child: Text(
-            //             "Radar Pulse",
-            //             style: TextStyle(color: Colors.white),
-            //           ),
-            //         ),
-            //       )
-            //     ],
-            //   ),
-            // )
+                      width: 100,
+                      height: 100,
+                      alignment: Alignment.center,
+                      decoration: BoxDecoration(
+                          color: Colors.white,
+                          borderRadius: BorderRadius.circular(60)),
+                      child: Text(
+                        "Snake Border",
+                        style: TextStyle(color: Colors.black),
+                      ),
+                    ),
+                  )
+                ]))
           ],
         ),
       ),