+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+
+## 0.0.4
+
+* Added Snake border .
+* Resolved flickering issue in gradient border.
+* Converted clipPath to custom painter for better performance.
+
 ## 0.0.3
 
 * Added Pulsating border.